% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BarPlot.R
\name{bar_plot}
\alias{bar_plot}
\title{Bar Plot Generator}
\usage{
bar_plot(dataset, group, value, y, ylimit)
}
\arguments{
\item{dataset}{-- Data Frame to evaluate}

\item{group}{-- Column in data (the variable) that you wish to look at.}

\item{value}{--  Is value in the column ( the variable) you wish to create a density plot}

\item{plot}{--  the parameter that you wish to use for the density plot}
}
