% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/choose.R
\name{choose}
\alias{choose}
\title{Returns one or n random elements of your list or dataset}
\usage{
choose(x = NULL, n = 1, replace = FALSE)
}
\arguments{
\item{x}{dataset (array or data frame) where to choose from}

\item{n}{number of choices (sample size)}

\item{replace}{sample with replacement or not. Default is FALSE.}
}
\value{

}
\description{
This function allows you to choose easily random elements of your dataset.
}
\examples{
# Choose a random number between 1 and 3
choose(c(1,2,3))
}
