% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CIplot.R
\name{CIplot}
\alias{CIplot}
\title{Generates a confidence limit visualization}
\usage{
CIplot(sample_size = 100, cl = 0.95)
}
\arguments{
\item{sample_size}{the number of obs in each sample}

\item{cl}{confidence limit}
}
