% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/collapse.R
\name{collapse}
\alias{collapse}
\title{Collapse Elements into a String}
\usage{
collapse(..., collapse = ", ")
}
\arguments{
\item{...}{Additional arguments which are to be collapsed into the output string.}

\item{collapse}{A character string to separate the elements of the output string. Defaults to a comma.}

\item{x}{An object or multiple objects (..., see below) to be collapsed into a single string.}
}
\value{
A single character string containing all input elements, separated by the 'collapse' character.
}
\description{
This function takes multiple arguments and collapses them into a single
character string, with a specified character used to separate the elements.
}
\examples{
collapse("Hello", "world", collapse = " ")
collapse("R", "is", "awesome", collapse = "_")
}
