% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/count_a_group.R
\name{count_a_group}
\alias{count_a_group}
\title{Create a function that provides the top n number of a grouped
variable by counts}
\usage{
count_a_group(data_name, groupings)
}
\arguments{
\item{data_name}{Name of data that need to be group}

\item{groupings}{Groups}
}
