% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DensityPlot.R
\name{density_plot}
\alias{density_plot}
\title{Density Plot Generator}
\usage{
density_plot(dataset, group, value, y, ylimit, bandwidth = 1)
}
\arguments{
\item{dataset}{-- Data Frame to evaluate}

\item{group}{-- Column in data (the variable) that you wish to look at.}

\item{value}{--  Is value in the column ( the variable) you wish to create a density plot}

\item{plot}{--  the parameter that you wish to use for the density plot}

\item{Bandwith}{-- width of smothting in kernal density estimation}
}
