% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/discard_all_empty_rows_cols.R
\name{discard_all_empty_columns}
\alias{discard_all_empty_columns}
\title{Discard All Empty Columns from a Data Frame}
\usage{
discard_all_empty_columns(d)
}
\arguments{
\item{d}{A data frame from which you want to remove columns.}
}
\value{
A data frame without the columns where all elements are empty.
}
\description{
This function removes all columns from a data frame where all elements of the column are empty.
}
\examples{
df <- data.frame(a = c(NA, NA, NA, NA),
                 b = c(1, NA, 5, 2))
discard_all_empty_columns(df)

}
