% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/discard_all_empty_rows_cols.R
\name{discard_all_empty_rows}
\alias{discard_all_empty_rows}
\title{Discard All Empty Rows from a Data Frame}
\usage{
discard_all_empty_rows(d)
}
\arguments{
\item{d}{A data frame that you want to filter.}
}
\value{
A data frame without the rows where all elements are empty.
}
\description{
This function removes all rows from a data frame where all elements of the row are empty.
}
\examples{
df <- data.frame(a = c(NA, 2, 3, NA),
                 b = c(NA, NA, 5, NA))
discard_all_empty_rows(df)

}
