% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/choose.R
\name{.choose_from_df}
\alias{.choose_from_df}
\title{Internal function - Samples data frames}
\usage{
.choose_from_df(x = x, n = n, replace = replace)
}
\arguments{
\item{x}{data frame to be sampled}

\item{n}{sample size}

\item{replace}{sample with replacement or not}
}
\value{
data frame
}
\description{
Internal function - Samples data frames
}
\keyword{internal}
