% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/choose.R
\name{.choose_from_list}
\alias{.choose_from_list}
\title{Internal function - Sampels lists}
\usage{
.choose_from_list(x = x, n = n, replace = replace)
}
\arguments{
\item{x}{list to be sampled}

\item{n}{sample size}

\item{replace}{sample with replacement or not}
}
\value{

}
\description{
Internal function - Sampels lists
}
\keyword{internal}
