% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocessing.R
\name{ds_ContToCat}
\alias{ds_ContToCat}
\title{Tranform some continuous column to categorical.}
\usage{
ds_ContToCat(dt, col_names, group_count = 5, inline = T)
}
\arguments{
\item{dt}{A data.table for tranform.}

\item{col_names}{A character vector columns name for transform.}

\item{group_count}{An integer, default 5. Quantity group.}

\item{inline}{A boolean, default TRUE. Modify current data.table or return new data.table.}
}
\description{
Tranform some continuous column to categorical.
}
