% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocessing.R
\name{ds_IntToFactor}
\alias{ds_IntToFactor}
\title{Transform int columns to factor.}
\usage{
ds_IntToFactor(dt, threshold = 10, inline = T, exclude = c())
}
\arguments{
\item{dt}{A data.table for modify.}

\item{threshold}{A integer, default 10. Number unique value in column for transform.}

\item{inline}{A boolean, default TRUE. Modify current data.table or return new data.table.}

\item{exclude}{A character vector for exclude columns.}
}
\description{
Function tranform integer columns in factor by threshold.
}
