% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocessing.R
\name{ds_MarkOutlier}
\alias{ds_MarkOutlier}
\title{Mark outliers in data.table.}
\usage{
ds_MarkOutlier(dt, one_column = T, inline = T, iqr_ratio = 1.5)
}
\arguments{
\item{dt}{A data.table for modify.}

\item{one_column}{A boolean, default TRUE. One or more out column.}

\item{inline}{A boolean, default TRUE. Modify current data.table or return new data.table.}

\item{iqr_ratio}{A numeric, default 1.5. IQR ratio for detect outlier}
}
\description{
Add column with name or prefix 'out' and mark every row: -1 normal, 1 outlier.
}
\examples{
ds_mark_outlier(train.data)
ds_mark_outlier(train.data, one_column = TRUE, inline = FALSE)
}
