% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocessing.R
\name{ds_toSparseMatrix}
\alias{ds_toSparseMatrix}
\title{Tranform data.table to sparse matrix}
\usage{
ds_toSparseMatrix(dt)
}
\arguments{
\item{dt}{A data.table for tranform.}
}
\description{
Tranform data.table to sparse matrix
}
