% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/entropy.R
\name{entropy}
\alias{entropy}
\title{Computes Shannon and Weaver's entropy}
\usage{
entropy(x = NULL, normalize = FALSE)
}
\arguments{
\item{x}{A vector for entropy computation}

\item{normalize}{Default is FALSE. Set \code{TRUE} to get normalized entropy.}
}
\value{
entropy value
}
\description{
Computes standard and normalized entropy based on Shannon and
Weavers' mathematical theory of communication (1948).
}
\examples{
entropy(x = c(0,0,1,1))
entropy(x = c(0,0,0,0))
}
