% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ext.R
\name{file_ext}
\alias{file_ext}
\title{Extract File Extension}
\usage{
file_ext(x)
}
\arguments{
\item{x}{A character string specifying the filename from which to extract the extension.}
}
\value{
The extension of the specified file as a character string. If the file has no extension, returns an empty string.
}
\description{
This function extracts the extension from a given filename.
}
\examples{
\dontrun{
# Extract the extension from a filename
ext <- file_ext('my_file.txt')
}
}
