% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flatten_list.R
\name{flatten_list}
\alias{flatten_list}
\title{Flatten a Nested List}
\usage{
flatten_list(lst)
}
\arguments{
\item{lst}{A list object, which can be a nested list (a list of lists).}
}
\value{
A flat list containing all elements from the input list. If the input list
  was nested, the returned list will be a single-level list combining all elements.
}
\description{
This function recursively traverses through a nested list (i.e., a list of lists)
and returns a single flat list containing all elements.
}
\examples{
nested_list <- list(a = 1, b = list(c = 2, d = 3, e = list(f = 4, g = 5)))
flatten_list(nested_list)
}
