% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Function_ideas.R
\name{Function_ideas}
\alias{Function_ideas}
\title{Generate sampling distribution(Function Ideas)}
\usage{
Function_ideas(samples, sample_size, mean, sd)
}
\arguments{
\item{samples}{Specity how many samples to run}

\item{sample_size}{the number of obs in each sample}

\item{mean}{population mean}

\item{sd}{population sd}
}
