% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataframeCompare.R
\name{fx_Align_Impute}
\alias{fx_Align_Impute}
\title{Align Attribute Classes, Factor Levels, and Impute Missing Values}
\usage{
fx_Align_Impute(original_dataframe, transform_dataframe,
  exclude_attribute_list, factor_imputation, numerical_imputation)
}
\arguments{
\item{original_dataframe}{The source dataframe to be referenced}

\item{transform_dataframe}{The dataframe to be transformed}

\item{exclude_attribute_list}{Optional list of attributes/columns to be ignore from both data.frames}

\item{factor_imputation}{Optional method for handling extra levels for factor attributes: "edf", "mode", "uniform", "NA", "ignore"}

\item{numerical_imputation}{Optional method for handling NAs for numerical attributes: "mean", "median", "mode", "edf", "ignore"}
}
\value{
A modified data.frame that resembles the original data.frame with aligned classes, factor levels, and
imputed values for NA elements
}
\description{
This function takes a reference data.frame and alters a target data.frame to resemble the reference. The
classes, factor levels, and imputed values will be taken from the reference data.frame unless otherwise
specified in the parameters.
}
