% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataframeCompare.R
\name{fx_allClass}
\alias{fx_allClass}
\title{List Feature Classes}
\usage{
fx_allClass(x)
}
\arguments{
\item{x}{The data.frame to be evaluated}
}
\value{
The classes of the data.frame columns
}
\description{
This function returns a character vector of the classes of the columns
of a given data.frame
}
