% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataframeCompare.R
\name{fx_mode}
\alias{fx_mode}
\title{Find Mode of a Vector}
\usage{
fx_mode(x)
}
\arguments{
\item{x}{The vector of values to reference}
}
\value{
The most frequently occuring value
}
\description{
This function finds the most frequently occuring element within a
vector. This will remove any NA elements before calculation.
}
