% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getfun.R
\name{getfun}
\alias{getfun}
\title{Retrieve an Exported Function from a Namespace}
\usage{
getfun(x)
}
\arguments{
\item{x}{A character string. If it includes "::", the string should be in the
format "namespace::function_name". If it does not include "::", the function
returns 'x' as it is.}
}
\value{
If 'x' includes "::", the function returns the function specified by
  'x' from the namespace specified by 'x'. If 'x' does not include "::", the
  function returns 'x' as it is.
}
\description{
This function retrieves an exported function from a specific namespace.
If the provided string does not include a namespace (i.e., does not contain "::"),
the function returns the string as it is.
}
\examples{
getfun("base::sum")
getfun("hello")
}
