% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggDensity.R
\name{ggDensity}
\alias{ggDensity}
\title{generates a density plot for a specific variable in a data.frame}
\usage{
ggDensity(df = mtcars, filter.var = "cyl", filter.val = 6,
  variable = "wt")
}
\arguments{
\item{df}{data.frame}
}
