% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/na_rows.R
\name{has_all_na_rows}
\alias{has_all_na_rows}
\title{Check for Rows with All NA}
\usage{
has_all_na_rows(d, cols = NULL)
}
\arguments{
\item{d}{A data frame to check for rows with all NA values.}

\item{cols}{An optional character vector specifying the columns to check for NA values.
If NULL (the default), all columns are considered.}
}
\value{
A logical value indicating whether any row with all NA values was found in the specified columns.
}
\description{
This function checks a given dataframe for rows where all values in the specified columns (or all columns if none are specified) are NA.
}
\examples{
\dontrun{
df <- data.frame(a = c(1, NA, 3), b = c(NA, NA, 3))
has_all_na_rows(df)
}
}
