% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/information_gain.R
\name{information_gain}
\alias{information_gain}
\title{Computes information gain based on Shannon and Weaver's entropy}
\usage{
information_gain(data, feature, target, bins = 4)
}
\arguments{
\item{data}{Data frame for information gain computation}

\item{feature}{Selected input feature}

\item{target}{Selected target feature}

\item{bins}{Number of bins to handle numeric features}
}
\value{
information gain value
}
\description{
Computes information gain based on standard entropy.
}
\examples{
information_gain(whisky_collection, "RATING", "PRICE", bins=5)
}
