% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is.empty.R
\name{is.empty}
\alias{is.empty}
\title{Checking emptiness}
\usage{
is.empty(x)
}
\arguments{
\item{x}{an objetc}
}
\value{
TRUE if the object is empty. FALSE is the object is not empty
}
\description{
Checking emptiness
}
\examples{

empty_object <- list()
is.empty(empty_object)

non_empty_object <- "hello"
is.empty(non_empty_object)

}
