% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_url.R
\name{is_url}
\alias{is_url}
\title{This function checks whether a given string is a URL. It recognizes both
'http://' and 'https://' protocols.}
\usage{
is_url(x)
}
\arguments{
\item{x}{A character string.}
}
\value{
A logical value: TRUE if 'x' is a URL (starts with 'http://' or
  'https://'), and FALSE otherwise.
}
\description{
This function checks whether a given string is a URL. It recognizes both
'http://' and 'https://' protocols.
}
\examples{
is_url("http://example.com")
is_url("https://example.com")
is_url("not_a_url")
}
