% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/JitterPlot.R
\name{jitter_plot}
\alias{jitter_plot}
\title{Bar Plot Generator}
\usage{
jitter_plot(dataset, group, value = "", x, y)
}
\arguments{
\item{dataset}{-- Data Frame to evaluate}

\item{group}{-- Column in data (the variable) that you wish to look at.}

\item{value}{--  Is value in the column ( the variable) you wish to create a density plot}

\item{plot}{--  the parameter that you wish to use for the density plot}
}
