% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list_pals.R
\name{list_pals}
\alias{list_pals}
\title{List the available palettes in the palette collection}
\usage{
list_pals()
}
\value{
list of colour palettes
}
\description{
This command lists you all the available color palettes in the dstools color
palette collection. The collection consists of a large collection of famous
colour palettes from popular brands and topics. This allows
you easily to adjust your \code{ggplot} to your design.
}
\examples{
list_pals()
}
