% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/named_list_to_df.R
\name{named_list_to_df}
\alias{named_list_to_df}
\title{Convert a Named List to a Data Frame}
\usage{
named_list_to_df(x, keep_null = TRUE, class = FALSE)
}
\arguments{
\item{x}{A named list.}
}
\value{
A tibble with two columns: 'name' (containing the names of the list elements)
  and 'value' (containing the values of the list elements).
}
\description{
This function takes a named list as input and returns a two-column data frame (or tibble).
The first column contains the names of the list elements, and the second column contains
the values of the list elements.
}
\examples{
named_list <- list("a" = 1, "b" = 2, "c" = 3)
named_list_to_df(named_list)
}
