% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalize.R
\name{normalize}
\alias{normalize}
\title{Use this function to normalize your data to a specific range}
\usage{
normalize(x)
}
\arguments{
\item{x}{A vector with numeric values}
}
\value{
A vector with min-max normalized values
}
\description{
Normalize your data
}
\examples{
data(whisky_collection)
whisky_collection$Rating <- normalize(whisky_collection$Rating)
}
