% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pipe.R
\name{\%||\%}
\alias{\%||\%}
\title{NULL Pipe}
\usage{
x \%||\% y
}
\arguments{
\item{x}{any value}

\item{y}{any value}
}
\value{
Return y if x is NULL, empty or NA
}
\description{
Pipe `%||%` evaluates nullity between two objects and:
a) **return x** if neither x nor y is NULL, empty or NA
b) **return x** if y is NULL, empty or NA
c) **return y** if x is NULL, empty or NA
}
\examples{

x <- "pepe"
y <- "x"

x \%||\% y


}
