% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/productionlog_sample.R
\docType{data}
\name{productionlog_sample}
\alias{productionlog_sample}
\title{Dataset with a sample of the Junglivet production data}
\format{
A data.frame with 21 rows and 8 variables:
\describe{
  \item{DAY}{day of the production}
  \item{MONTH}{month of the production}
  \item{MANUFACTURER}{name of the leading brew master)}
  \item{PRODUCT}{product type like Junglivet or Junglivet premium}
  \item{SHIFT}{indicates morning or evening shift)}
  \item{COLOR}{color indicator on a scale between 0 and 1}
  \item{MALTING}{origin of the used malting products}
  \item{TASTING}{quality indicator based on a whisky tasting}
}
}
\usage{
data(productionlog_sample)
}
\description{
A sample of the production data of the Junglivet whisky
}
\examples{
data(productionlog_sample)
\donttest{
library("ggplot2")
ggplot(productionlog_sample, aes(x=MANUFACTURER, y=TASTING, fill=TASTING)) +
geom_bar(stat="identity") +
coord_flip() +
xlab("")
}
}
\keyword{datasets}
