% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/QQPlot.R
\name{qq_plot}
\alias{qq_plot}
\title{QQ Plot Generator}
\usage{
qq_plot(dataset, group, value, plot)
}
\arguments{
\item{dataset}{-- Data Frame to evaluate}

\item{group}{-- Column in data (the variable) that you wish to look at.}

\item{value}{--  Is value in the column ( the variable) you wish to create a density plot}

\item{plot}{--  the parameter that you wish to use for the qq plot}
}
