% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/random_name.R
\name{random_name}
\alias{random_name}
\title{Generate a Random Name}
\usage{
random_name(n = 8, prefix = "")
}
\arguments{
\item{n}{An integer specifying the length of the random part of the name.
Defaults to 8.}

\item{prefix}{An optional character string to be appended to the start
of the random name. Defaults to an empty string.}
}
\value{
A character string consisting of the prefix (if any) followed by a
  random alphanumeric string.
}
\description{
This function generates a random alphanumeric string of a specified length,
with an optional prefix. The output can be used to create random identifiers
or temporary names.
}
\examples{
random_name()
random_name(5)
random_name(10, "temp_")
}
