% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_dir.R
\name{read_dir}
\alias{read_dir}
\title{Wrapper for different \code{read functions}.}
\usage{
read_dir(dir = NULL, pattern = NULL, type = c("xlsx"))
}
\arguments{
\item{dir}{Path to the directory of files}

\item{pattern}{Specify a word or name to read only
files with this specific word in the filename and in the directory}

\item{type}{Choose filetype (e.g. xlsx) that should be read}
}
\value{
A data.frame consisting of all related files
}
\description{
Reads all files in a specific directory into your
environment and binds them into a data.frame.
}
\examples{
\dontrun{data_all <- read_dir(dir = C:/users/Username/data, type = "xlsx")}
\dontrun{data_all <- read_dir(dir = C:/logs, pattern = "logs", type = "xlsx")}
}
