% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remove_accents.R
\name{remove_accents}
\alias{remove_accents}
\title{Remove accents}
\usage{
remove_accents(string)
}
\arguments{
\item{string}{a character with accents}
}
\value{
a character without accents
}
\description{
Remove accents
}
\examples{

remove_accents("Estós ácentós están por sér elíminádós")
}
