% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/removeNulls.R
\name{removeNulls}
\alias{removeNulls}
\title{Remove NULL values}
\usage{
removeNulls(x)
}
\arguments{
\item{x}{vector with NULL value(s)}
}
\value{
vector without NULL values
}
\description{
Remove NULL values
}
