% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rename_columns.R
\name{rename_columns}
\alias{rename_columns}
\title{Rename data frame columns using a dictionary}
\usage{
rename_columns(df, name_dict)
}
\arguments{
\item{df}{A data frame whose columns you want to rename.}

\item{name_dict}{A named vector where names are the current column names and values
are the new column names, or a data frame with two columns where the first column
contains new names and the second column contains original names.}
}
\value{
A data frame with renamed columns.
}
\description{
This function renames the columns of a data frame based on a provided dictionary.
It only renames columns that exist in the data frame and ignores dictionary
entries for non-existent columns. The dictionary can be either a named vector
or a data frame with two columns. The first column will be the new name if the
input is a data.frame. The new name will be the names of the input dictionary.
}
\examples{
df <- data.frame(a = 1:3, b = 4:6, c = 7:9)

# Using a named vector
vec_dict <- c(a = "A", b = "B", d = "D")
rename_columns(df, vec_dict)

# Using a data frame
df_dict <- data.frame(new_name = c("A", "B", "D"),
                      original_name = c("a", "b", "d"))
rename_columns(df, df_dict)

}
