% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SamplingDist.R
\name{SamplingDist}
\alias{SamplingDist}
\title{Generates a poor man's sampling disribution}
\usage{
SamplingDist(samples = 50, sample_size = 100)
}
\arguments{
\item{samples:}{Specifcy how many samples to run}

\item{sample_size:}{the number of obs in each sample}
}
