% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ext.R
\name{sans_ext}
\alias{sans_ext}
\title{Remove File Extension}
\usage{
sans_ext(x, compression = FALSE)
}
\arguments{
\item{x}{A character string specifying the filename from which to remove the extension.}

\item{compression}{A logical value indicating whether to remove compression extensions (gz, bz2, xz).
Default is FALSE.}
}
\value{
The filename without its extension as a character string.
}
\description{
This function removes the extension from a given filename. If the compression parameter is TRUE,
it additionally removes the compression extension (gz, bz2, or xz).
}
\examples{
\dontrun{
# Remove the extension from a filename
filename <- sans_ext('my_file.txt.gz', compression = TRUE)
}
}
