% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Scatter_Faucet_Plot.R
\name{scatter_fuacet_plot}
\alias{scatter_fuacet_plot}
\title{Top Data by Count (returns new data set with only top values of choosen group)}
\usage{
scatter_fuacet_plot(dataset, group, x, n, limit, bw = 1)
}
\arguments{
\item{dataset}{-- Data Frame to evaluate}

\item{group}{-- Grouping that you wish to use on the data frame}

\item{n}{--  Is the number of top items in the group you want to display}
}
