% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Scatter_Histo_Plot.R
\name{scatter_histo_plot}
\alias{scatter_histo_plot}
\title{Scatter Plot with histograms Generator}
\usage{
scatter_histo_plot(dataset, group, value = "", x, y, ylimit)
}
\arguments{
\item{dataset}{-- Data Frame to evaluate}

\item{group}{--    Column in data (the variable) that you wish to look at.}

\item{value}{--    Is value in the column ( the variable) you wish to create a plot for}

\item{x}{--     The control or input variable}

\item{y}{--     The result or output variable that you wish to use for plotting}

\item{ylimit}{-- limit on the y value}
}
