% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ScatterPlot.R
\name{scatter_plot}
\alias{scatter_plot}
\title{Scatter Plot Generator}
\usage{
scatter_plot(dataset, group, value, plot, ylimit)
}
\arguments{
\item{dataset}{-- Data Frame to evaluate}

\item{group}{-- Column in data (the variable) that you wish to look at.}

\item{value}{--  Is value in the column ( the variable) you wish to create a density plot}

\item{plot}{--  the parameter that you wish to use for the qq plot}
}
