% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statmode.R
\name{statmode}
\alias{statmode}
\title{Use this function to compute the statistical mode of a dataset}
\usage{
statmode(x, na.rm = FALSE)
}
\arguments{
\item{x}{A vector with numeric or character values}

\item{na.rm}{Specify if NA should be removed}
}
\value{
A value with the statistical mode
}
\description{
Compute the mode
}
\examples{
data(whisky_collection)
statmode(whisky_collection$RATING)
}
