% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StatsPlots.R
\name{stats_plot}
\alias{stats_plot}
\title{Statistic Graphs Comparison}
\usage{
stats_plot(dataset, group, value, plot, bandwidth = 1)
}
\arguments{
\item{dataset}{--  Data Frame to evaluate}

\item{group}{--    Column in data (the variable) that you wish to look at.}

\item{value}{--    Is value in the column ( the variable) you wish to create a plot for}

\item{plot}{--     The parameter that you wish to use for plotting}

\item{Bandwith}{-- width of smothting in kernal density estimation}
}
