% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/supermarket.R
\docType{data}
\name{supermarket}
\alias{supermarket}
\title{Dataset with a transaction log sample from the Junglivet onlineshop}
\format{
A transaction set with 9835 transactions aggregated to 170 categories:
\describe{
  \item{item}{name of the item in the transaction}
 }
}
\usage{
data(supermarket)
}
\description{
Transaction logs of a Scottish supermarket including Junglivet whisky sales
based on the dataset from Michael Hahsler et al. (2006).
}
\details{
Source: Michael Hahsler, Kurt Hornik, and Thomas Reutterer (2006)
Implications of probabilistic data modeling for mining association rules.
In M. Spiliopoulou, R. Kruse, C. Borgelt, A. Nuernberger, and W. Gaul,
editors, From Data and Information Analysis to Knowledge Engineering,
Studies in Classification, Data Analysis, and Knowledge Organization,
pages 598–605. Springer-Verlag.
}
\examples{
data(supermarket)
\donttest{
summary(supermarket)
}
}
\keyword{datasets}
