% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TopGroup.R
\name{top_group}
\alias{top_group}
\title{Top Grouping by Count}
\usage{
top_group(dataset, group, n)
}
\arguments{
\item{dataset}{-- Data Frame to evaluate}

\item{group}{-- Grouping that you wish to use on the data frame}

\item{n}{--  Is the number of top items in the group you want to display}
}
