% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unix_timestamp.R
\name{unix_timestamp}
\alias{unix_timestamp}
\title{Convert the current system time to a UNIX timestamp}
\usage{
unix_timestamp()
}
\value{
An integer, representing the current UNIX timestamp.
}
\description{
This function returns the current system time as a UNIX timestamp.
A UNIX timestamp represents the number of seconds since '1970-01-01 00:00:00 UTC'.
}
