% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unix_timestamp.R
\name{unix_timestamp_to_date}
\alias{unix_timestamp_to_date}
\title{Convert a UNIX timestamp to a POSIXct date-time}
\usage{
unix_timestamp_to_date(x)
}
\arguments{
\item{x}{An integer or numeric vector representing a UNIX timestamp.}
}
\value{
A POSIXct object representing the converted date-time.
}
\description{
This function converts a provided UNIX timestamp into a POSIXct date-time object.
A UNIX timestamp represents the number of seconds since '1970-01-01 00:00:00 UTC'.
}
