% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/na_rows.R
\name{which_all_na_rows}
\alias{which_all_na_rows}
\title{Identify Rows with All NA}
\usage{
which_all_na_rows(d, cols = NULL, na = c(NA))
}
\arguments{
\item{d}{A data frame to check for rows with all NA values.}

\item{cols}{An optional character vector specifying the columns to check for NA values.
If NULL (the default), all columns are considered.}

\item{na}{An optional vector specifying the values to be considered as NA. Default is c(NA).}
}
\value{
A vector of indices of rows with all NA values in the specified columns.
}
\description{
This function identifies the indices of rows in a given dataframe where all values in the specified columns
(or all columns if none are specified) are NA.
}
\examples{
\dontrun{
df <- data.frame(a = c(1, NA, 3), b = c(NA, NA, 3))
which_all_na_rows(df)
}
}
