% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/which_not_in.R
\name{which_not_in}
\alias{which_not_in}
\title{Find which values are in one vector and not in the other}
\usage{
which_not_in(x, y)
}
\arguments{
\item{x}{character vector}

\item{y}{character vector to check for existance}
}
\value{
The values in x not in y
}
\description{
Find which values are in one vector and not in the other
}
\examples{

x <- c("a","b", "missing")
y <- c("a", "b")

which_not_in(sample, "ex_na")

}
