% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/whiskycollection.R
\docType{data}
\name{whisky_collection}
\alias{whisky_collection}
\title{Dataset with a collection of popular whiskies}
\format{
A data.frame with 36 rows and 11 variables:
\describe{
  \item{Name}{Name of the whisky}
  \item{Distillery}{Distiller of the specific whisky}
  \item{Location}{Production location of the whisky (mostly countries or regions)}
  \item{Type}{Specification of the whisky type like e.g. single malt or blended}
  \item{Region}{Region of the whisky production (mostly relevant for scotchs)}
  \item{Foundation}{Year of the first whisky production}
  \item{Coordinates}{Longitude and latitude values of the destillery}
  \item{Wikipedia}{Link to the related article of the english wikipedia}
  \item{Rating}{My personal rating of this whisky. I am open to discuss it, just write me an email if you see it otherwise}
  \item{Smokness}{My measure of how smoky vs. delicate it tastes, negative values implicate delicate}
  \item{Richness}{My measure of how rich vs. light it tastes, negative values implicate light}
}
}
\usage{
data(whisky_collection)
}
\description{
Data about popular whiskies around the world with additional information
}
\examples{
data(whisky_collection)
\donttest{
library("ggplot2")
ggplot(whisky_collection, aes(x=Name, y=Rating, fill=Rating)) +
geom_bar(stat="identity") +
coord_flip() +
xlab("")
}
}
\keyword{datasets}
