% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/whiskycollection.R
\docType{data}
\name{whiskycollection}
\alias{whiskycollection}
\title{Arabidopsis QTL data on gravitropism}
\format{
A data frame with NNNN rows and NN variables:
\describe{
  \item{Name}{Anonymized Mechanical Turk Worker ID}
  \item{Rating}{Trial number, from 1..NNN}
  ...
}
}
\source{
\href{https://phenome.jax.org/projects/Moore1b}{QTL Archive}
}
\usage{
data(whiskycollection)
}
\description{
Data from a QTL experiment on gravitropism in
Arabidopsis, with data on 162 recombinant inbred lines (Ler x
Cvi). The outcome is the root tip angle (in degrees) at two-minute
increments over eight hours.
}
\examples{
data(whiskycollection)
\donttest{plot(whiskycollection)}
}
\references{
Moore et al. (2013) Genetics 195:1077-1086
(\href{https://www.ncbi.nlm.nih.gov/pubmed/23979570}{PubMed})
}
\keyword{datasets}
