% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DensityPlot.R
\name{density_plot}
\alias{density_plot}
\title{Density Plot Generator}
\usage{
density_plot(dataset, col_name, rank, value = "")
}
\arguments{
\item{dataset}{-- Data Frame to evaluate}

\item{col_name}{-- Column in data that you wish to look at.}

\item{rank}{--  the parameter that you wish to use for the density plot}

\item{value}{--  Is value in the column you wish to look at}
}
