% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggdotplot.R
\name{ggdotplot}
\alias{ggdotplot}
\title{Creates a Dotplot}
\usage{
ggdotplot(x, binwidth = NULL, dotsize = 0.75)
}
\arguments{
\item{x}{A vector of numeric values}

\item{binwidth}{The width of a column of dots}

\item{dotsize}{The relative size of a dot. Smaller mean less dot overlap.}
}
\description{
The scaling for ggplot2::dotplot() is unfortunately designed
for more complex situations than what is used in an introductory
resampling-based statistics course. It would be very nice to
be able to make a geom_dotplot() function that respected the
y-axis.
}
\examples{
ggdotplot( rnorm(100) )

}
