% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pipe.R
\name{\%||\%}
\alias{\%||\%}
\title{NULL Pipe}
\usage{
x \%||\% y
}
\arguments{
\item{x}{any value}

\item{y}{any value}
}
\value{
Return y if x is NULL, empty or NA
}
\description{
Defaults for NULL values
}
\examples{

x <- "pepe"
y <- "x"

x \%||\% y


}
