% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SamplingDist.R
\name{SamplingDist}
\alias{SamplingDist}
\title{Sampling Distribution Function for Fun}
\usage{
SamplingDist(samples = 50, sample_size = 1000, mean = 0, sd = 1)
}
\arguments{
\item{sample_size}{Specifies the size of the sample you are taking}

\item{mean}{of the normal function for generating the individual random points}

\item{sd}{the standard deviation of the normal function for generating the individual random points.}

\item{Samples}{Specify how many samples that you will do}
}
