% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SamplingDist.R
\name{SamplingDist}
\alias{SamplingDist}
\title{Sampling Distribution Function for Fun}
\usage{
SamplingDist(samples = 50, sample_size = 1000, mean = 0, sd = 1)
}
\arguments{
\item{Samples}{Specify how many samples}
}
