% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/whiskycollection.R
\docType{data}
\name{whiskycollection}
\alias{whiskycollection}
\title{Collection of popular whiskies}
\format{
A data frame with 3 rows and 2 variables:
\describe{
  \item{Name}{Name of the whisky}
  \item{Rating}{My personal rating, if available}
  ...
}
}
\usage{
data(whiskycollection)
}
\description{
Data about popular whiskies around the world
}
\examples{
data(whiskycollection)
\donttest{plot(whiskycollection)}
}
\keyword{datasets}
